refactor: move emitModel function
